+++ NEWS.md
@@ -15,6 +15,8 @@
 
 * Remove `qs` since it will be archived and replace with `qs2`.
 
+* Added `saemix` estimation method
+
 # babelmixr2 0.1.10
 
 * Bug fix for the new version of `units` (#179)